text-reader-app/backend/requirements.txt
fastapi==0.115.4
langchain_community==0.3.4
langchain_core==0.3.14
langchain_openai==0.2.4
langchain_text_splitters==0.3.1
httpx==0.27.2
openai==1.56.1
pandas==2.2.3
pydantic==2.9.2
python-dotenv==1.0.1
SQLAlchemy==2.0.34
uvicorn==0.30.6
psycopg2-binary==2.9.9
llama-parse==0.5.5
python-multipart==0.0.9